feat: add database functionality and state management
I did not fully understand the purpose of BooksAPI.js and realized that the data was already provided. I modified the initial state of the book list and added the features to get and to update the state of the book. I also added a preload view as well as a updating status view.